+++ src/wrapped_radix_engine_toolkit/low_level/address_conversion/mod.rs
@@ -1,0 +1,2 @@
+mod resource_address_from;
+pub use resource_address_from::*;
